Initial Coq backend.
This adds --backend option coq
which generates Coq files defining the sorts and rules of a definition.

This first generates a file containing a parsed version of the rules of
the definition, and then runs a separate "kcoq" binary which generated
the actual Coq files by reading rule bodies from the generated file and
reading syntax from the actual .k files. The kcoq binaries are not
included in this commit.