Move Coq backend to separate maven module
This commit also modifies test code in kernel because some
unit tests used on the CoqBackend, apparently just as a convenient
Backend instance to test injection.
The dummy TestBackend class is added as a replacement, and the test
setup is modified to register it under the name "test".

imp.labeled_rules was regenerated with the parent of this commit,
and is modifed only because of other changes in the K framework.
Extracting the Coq backend to a module did not change it further.
The Coq file was not tested or updated, because it's made by the
external kcoq program instead of the Java Coq Backend code.