Move developer instructions to repo root
After this commit, developer instructions for working from source are in
the repo root, user instructions meant for a release zip are in the
k-distribution directory, and the maven assembly file makes sure it's
still the user instructions that end up at the root of k directory in a
release zip.

This does not change the content of the files, besides adjusting
references to the developer documentation, and updating a copyright header.