Refactor parseString to return K.
All callers immediately applied TreeNodesToKORE.apply,
so it's convenient to add it into parseString.

Also, a recent change to AmbFilter makes the result somewhat
unpredictable as a Term by picking an arbitrary resolution of
any ambiguity where the unequal Terms become equivalent
after applying TreeNodesToKORE.apply, so it's safer not
to expose the Term-returning version.