Fix configuration generation for optional cells.
A previous commit stopped generating the "unit" attribute
which later passes used to identify optional cells, especially
affecting configuration abstraction.
This always produces compiler errors for non-initial cells,
but only for intial cells if some other rule writes a closed
parent cell without the optional cell.
(the generated intializer for a non-initial cell contains
a cell like that, always producing the error).

This was missed by the unit test because
GenerateSentencesFromConfigDeclTest was relying on Production.equals
which intentionally ignores attributes other than klabel.
The test is improved to additionally check attributes,
and the reference output updated with expected attributes.