Support cell rewrites.
The SortCells pass does not recurse to find cell rewrites hidden under
labels such as "#cells", and was thus incorrectly computing the sets of
neighboring cells of some cell fragment variables.
This commit fixes that. It also simplifies the logic by adding
a small pass to lift any rewrite inside a non-leaf cell to occur
as an immediate child of the parent cell.