Filter argument sorts
AddEmptyList now computes the sort of terms in labeled syntax more carefully.
The second argument of the #KApply is divided into arguments by
splitting on #KList applications.
If none have sort KList, it is assumed to be a list of concrete
arguments, and the productions for the klabel are filtered down
to those consistent with the arguments, before returning
the least sort among the remaining results.