Handle cell fragment labels in SortCells
Besides the changes in SortCells, this also needs to
add some information fo ConfigurationInfo.
This also reverts earlier changes to again pass a KExceptionManager
throught to SortCells, and use it to report some errors.

The kool-untyped example is updated to again use cell fragments,
and a regression test is added.
(these parts were done by Daejun)